+++ internal/registry/migrations/0001_signing_keys.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE axto_signing_keys;